+++ app/src/main/java/com/example/buggyweather/BuggyWeatherApp.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import com.example.buggyweather.home.HomeModule
 import com.example.buggyweather.main.MainModule
 import com.example.buggyweather.network.NetworkModule
-import com.example.buggyweather.preference.SharedPreferenceModule
+import com.example.buggyweather.core.preference.SharedPreferenceModule
 import com.example.buggyweather.whole.WholeDayModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
@@ -21,10 +21,11 @@ class BuggyWeatherApp : Application() {
 			modules(
 					NetworkModule.provide(),
 					MainModule.provide(),
-					SharedPreferenceModule.provide(),
 					HomeModule.provide(),
 					WholeDayModule.provide()
 			)
 		}
+
+		SharedPreferenceModule.init()
 	}
 }
