+++ core/src/main/java/com/example/buggyweather/core/preference/SharedPreferenceModule.kt
@@ -1,8 +1,9 @@
-package com.example.buggyweather.preference
+package com.example.buggyweather.core.preference
 
 import android.content.Context
 import android.content.SharedPreferences
 import com.example.buggyweather.core.base.KoinModule
+import org.koin.core.context.loadKoinModules
 import org.koin.core.module.Module
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
@@ -18,4 +19,6 @@ object SharedPreferenceModule : KoinModule {
 	override fun provide(): Module = module {
 		single(named(FEATURE_SETTING)) { createSharedPreference(get(), FEATURE_SETTING) }
 	}
+
+	fun init() = loadKoinModules(provide())
 }
